Added support for load balancing across multiple servers
It was somewhat tricky because it was important that people in the same room end up on the server. Basically how it works is, for the http part it doesn't matter which server users connect to. For the websocket part the client makes a hash of the room uid, does modulus nr or of servers and connects the websocket to that server.

It's now pretty messy to set up, with hardcoded ip addres in the code. I'll clean it up when the dust has settled a little.